# Gh 91 chart load performance