--- a/gulpfile.babel.ts
+++ b/gulpfile.babel.ts
@@ -153,6 +153,7 @@ export {
   postinstallCi,
   verifyCi,
   verify,
+  runCypressTests,
 };
 
 export default verify;