--- a/src/client/cypress.json
+++ b/src/client/cypress.json
@@ -1,7 +1,5 @@
 {
   "darkMediaQuery": true,
-  "defaultCommandTimeout": 100000,
-  "requestTimeout": 16000,
   "viewportHeight": 500,
   "viewportWidth": 1600,
   "watchForFileChanges": true,