--- a/src/client/cypress.json
+++ b/src/client/cypress.json
@@ -1,5 +1,7 @@
 {
   "darkMediaQuery": true,
+  "defaultCommandTimeout": 6000,
+  "requestTimeout": 7000,
   "viewportHeight": 500,
   "viewportWidth": 1600,
   "watchForFileChanges": true,