--- a/src/client/src/screens/App/index.tsx
+++ b/src/client/src/screens/App/index.tsx
@@ -1,4 +1,8 @@
-import ApolloClient from 'apollo-boost';
+import { InMemoryCache } from 'apollo-cache-inmemory';
+import { ApolloClient } from 'apollo-client';
+import { ApolloLink } from 'apollo-link';
+import { onError } from 'apollo-link-error';
+import { HttpLink } from 'apollo-link-http';
 import 'normalize.css';
 import React from 'react';
 import { ApolloProvider } from 'react-apollo';
@@ -21,7 +25,23 @@ if (serverUrl.includes('undefined')) {
 }
 
 const client = new ApolloClient({
-  uri: serverUrl,
+  link: ApolloLink.from([
+    onError(({ graphQLErrors, networkError }) => {
+      if (graphQLErrors) {
+        graphQLErrors.map(({ message, locations, path }) =>
+          console.log(
+            `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`,
+          ),
+        );
+      }
+      if (networkError) console.log(`[Network error]: ${networkError}`);
+    }),
+    new HttpLink({
+      uri: serverUrl,
+      credentials: 'same-origin',
+    }),
+  ]),
+  cache: new InMemoryCache(),
 });
 
 const AppStyles = styled.section`