--- a/src/client/src/screens/App/index.tsx
+++ b/src/client/src/screens/App/index.tsx
@@ -39,7 +39,6 @@ const client = new ApolloClient({
     new BatchHttpLink({
       uri: serverUrl,
       credentials: 'same-origin',
-      batchInterval: 25,
     }),
   ]),
   cache: new InMemoryCache(),