… src/client/package.json
… src/server/config/templates/docker-compose.yml