# Proposed changes

This resolves the horribly slow loading times on Euclid, which made it completely unusable from a UX perspective, and from a practical one. It now loads in under 3 seconds, which is a decent benchmark for UX loading times: since [40% of people will abandon a site that takes more than 3 seconds to load.](https://neilpatel.com/blog/loading-time/). Previously it loaded in about 15 to 20 seconds!!!! 🐌🐌🐌

* **What is the current behavior?** (You can also link to an open issue here)
Euclid is incredibly slow; takes 15+ seconds to load! 🐌🐌🐌🐌🐌🐌🐌 See #91 for more details.

* **What is the new behavior (if this is a feature change)?**
Euclid is very fast; takes <3 seconds to load!!! 🏎🏎🏎🏎🏎🏎🏎

## Types of changes

What types of changes does your code introduce to Euclid?

* [x] Bug-fix (non-breaking change which fixes an issue)
* [ ] New feature (non-breaking change which adds functionality)
* [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist

* [x] I have read the [CONTRIBUTING](https://github.com/dbpiper/Euclid/blob/docs/CONTRIBUTING.md) guide
and all commit messages follow it
* [x] Lint and unit tests pass locally with my changes
* [x] End-to-end tests pass locally with my changes
* [x] I have added tests that prove my fix is effective or that my feature works
* [x] I have added necessary documentation (if appropriate)

### Further comments

You can see #91 for details on how slow it actually was, as I took many screenshots of detailing my profiling process. 

Here are the _new_ screenshots showing the massive speedup:

![chart integration test times (after memoization fix)](https://user-images.githubusercontent.com/1874789/56168755-a1427e00-5fcb-11e9-9fa8-45a7a40c0c02.png)
![e2e test times (after memoization fix)](https://user-images.githubusercontent.com/1874789/56168692-6e988580-5fcb-11e9-9e36-f057aeae8c79.png)